+++ tests/strings_runes/main.rs
@@ -6,7 +6,7 @@ fn main() {
     println!("{}", (*s.borrow().as_ref().unwrap()).len());
 
     let mut i = Rc::new(RefCell::new(Some(0)));
-    while (*i.borrow().as_ref().unwrap()) < (*s.borrow().as_ref().unwrap()).len() {
+    while ((*i.borrow().as_ref().unwrap()) as i32) < ((*s.borrow().as_ref().unwrap()).len() as i32) {
         print!("{} ", ({ let __s = (*s.borrow().as_ref().unwrap()).clone(); __s.as_bytes()[((*i.borrow().as_ref().unwrap())) as usize] }) as u8 as char);
         { let mut guard = i.borrow_mut(); *guard = Some(guard.as_ref().unwrap() + 1); }
     }
